Migrate metrics from Graphite to Prometheus
This change migrates metrics from Graphite to Prometheus.

Due to the points in the code where some StatsD metrics were computed,
and because only one function is used to increment both
StatsD and Prometheus counters, it was necessary to refactor some
of the legacy metric names to prevent counting a success or failure
multiple times.

If still needed, the values for the old Graphite metrics fail.total
and success.total should now be displayed on dashboards as the sum of
all fail.* (or success.*) metrics.

Bug: T359381
Change-Id: I5736a36839f2f14441044090f4208379fe19c4e6